// 0.generate_random_inventories/lib/types.ts
export interface ProductDimensions {
    width: number;
    height: number;
    depth: number;
    weight: number;
}

export interface ConvenienceStoreProduct {
    name: string;
    description: string;
    quantity: number;
    location: string;
    width: number;
    height: number;
    depth: number;
    weight: number;
    price: number;
}